+++ CHANGES.md
@@ -10,11 +10,14 @@
   4. Down:  Jumps to the current month.
   6. F5:  Refreshes the current view.
   7. Ctrl/Command + F:  Opens the search dialog.
+  8. Ctrl/Command + Left:  Moves to the previous year.
+  9. Ctrl/Command + Right:  Moves to the next year.
 - Added "created" support to events (set automatically if not available to the current date/time).
 - Added "organizerName" support to events (set automatically if not available to options > organizerName).
 - Added "organizerEmailAddress" support to events (set automatically if not available to options > organizerEmailAddress).
 - Added support to export events to the iCAL format.
 - Fixed events being in the wrong order when exported (to any format).
+- Added public methods "moveToPreviousYear()" and "moveToNextYear()" (with new on events).
 
 ### Version 0.5.2:
 - The text "All Day Event" is now "All-Day" (renamed all the settings to match).